requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.10.2
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
asttokens==2.4.1
attrs==25.3.0
azure-ai-documentintelligence==1.0.0
azure-core==1.30.1
azure-cosmos==4.7.0
azure-functions==1.21.3
azure-functions-durable==1.2.10
azure-storage-blob==12.20.0
certifi==2024.7.4
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.8
colorama==0.4.6
comm==0.2.2
cryptography==44.0.2
debugpy==1.8.5
decorator==5.1.1
distro==1.9.0
docx2pdf==0.1.8
et-xmlfile==1.1.0
executing==2.0.1
fastapi==0.115.5
frozenlist==1.5.0
furl==2.1.4
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.7
ipykernel==6.29.5
ipython==8.26.0
isodate==0.7.2
jedi==0.19.1
jiter==0.9.0
jupyter_client==8.6.2
jupyter_core==5.7.2
matplotlib-inline==0.1.7
multidict==6.2.0
nest-asyncio==1.6.0
numpy==2.0.1
openai==1.55.3
openpyxl==3.1.5
orderedmultidict==1.0.1
packaging==24.1
pandas==2.2.2
parso==0.8.4
pillow==11.1.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
propcache==0.3.0
psutil==6.0.0
pure_eval==0.2.3
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
pyodbc==5.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2024.1
pywin32==306
pyzmq==26.1.0
reportlab==4.3.1
requests==2.32.3
six==1.16.0
sniffio==1.3.1
stack-data==0.6.3
starlette==0.41.3
tenacity==8.2.3
tornado==6.4.1
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
uvicorn==0.32.1
wcwidth==0.2.13
yarl==1.18.3
